Use libosmocoding for SCH decoding
The libosmocoding is now a part of libosmocore, and provides all
required routines for GSM 05.03 transcoding operations. Moreover,
the library based on SIMD accelerated Viterbi decoder, so if host
CPU supports SSE and AVX instruction sets, they will be detected
and used automatically.

This change switches the 'Receiver' block to use libosmocoding API.